# api/openapi.yaml
openapi: 3.0.3
info:
  title: Duongöndro API
  version: "1"
  description: |
    The server stores sealed blobs, public keys and signed statements. It never
    receives a practice count, a round or a key, and it checks only sizes, glowie
    curve (NIST P-256) points and Ed25519 signatures over the exact bytes it receives
    (docs/crypto.md). Binary fields are standard base64 (`format: byte`).

    Contract first: edit this file, then `make gen`. The Android client is generated
    from it too.

    Every sign-in ends in a bearer token. An account is created only with a valid
    invitation; signing in never grants keys. The DEV build's POST /api/dev/session is
    registered outside this file, so a release build has no trace of it.
servers:
  - url: https://api.duongondro.app
paths:
...
  /api/invites:
    get:
      operationId: listInvites
      parameters: [{$ref: "#/components/parameters/Authorization"}]
      responses:
        "200":
          description: Invites this account created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/InviteList"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    post:
      operationId: createInvite
      description: |
        Stores a reusable invitation: the inviter's signed invite statement, the pin
        MAC the invitee checks, and auth, of which the server keeps only a hash.
      parameters: [{$ref: "#/components/parameters/Authorization"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/InviteInput"}
      responses:
        "201": {description: Created.}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  parameters:
    Authorization:
      name: Authorization
      in: header
      required: true
      description: Bearer <session token>
      schema: {type: string}
...
  responses:
    Unauthorized:
      description: Missing, unknown or revoked session token.
    BadRequest:
      description: The request is malformed or fails validation.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
    Conflict:
      description: Conflicts with the stored state.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: {type: string}
...
    InviteInput:
      type: object
      required: [id, auth, expiresAt, payload, signature, mac]
      properties:
        id: {type: string, minLength: 8, maxLength: 8}
        auth: {type: string, format: byte, description: "HKDF(secret, invite-auth), 32 bytes"}
        expiresAt: {type: string, format: date-time}
        payload: {type: string, format: byte, description: The signed invite statement}
        signature: {type: string, format: byte}
        mac: {type: string, format: byte, description: "HMAC(pin, inviterIdentityPk), 32 bytes"}
...
    InviteSummary:
      type: object
      required: [id, expiresAt, createdAt]
      properties:
        id: {type: string}
        expiresAt: {type: string, format: date-time}
        revokedAt: {type: string, format: date-time}
        createdAt: {type: string, format: date-time}
    InviteList:
      type: object
      required: [invites]
      properties:
        invites:
          type: array
          items: {$ref: "#/components/schemas/InviteSummary"}
...